+++ lib/vagrant/hosts/freebsd.rb
@@ -11,6 +11,11 @@ module Vagrant
         Util::Platform.freebsd?
       end
 
+      # Normal, mid-range precedence.
+      def self.precedence
+        5
+      end
+
       def initialize(*args)
         super